novas classes e dialog

+++ Flashstudy_Mobile/app/src/main/java/br/com/flashstudy/flashstudy_mobile/offline/dao/CronogramaDao.java
@@ -25,5 +25,5 @@ public interface CronogramaDao {
 
     //Procura o Cronograma de um Usuario
     @Query("SELECT * FROM cronograma WHERE usuario_codigo = :codigo")
-    CronogramaOff getCronogramaByUsuario(int codigo);
+    CronogramaOff getCronogramaByUsuario(long codigo);
 }
